review feedback + new db version
