+++ scheduler/requirements.txt
@@ -1,5 +1,5 @@
 celery
 python-dotenv
-git+https://github.com/PerfectFit-project/virtual-coach-db@v.0.1.12
+git+https://github.com/PerfectFit-project/virtual-coach-db@v0.1.13
 redis
 requests
